Moved request info and environment info collector logic to respective plugins
