--- src/main/java/com/exceptionless/exceptionlessclient/services/EnvironmentInfoCollectorIF.java
@@ -1,7 +1,0 @@
-package com.exceptionless.exceptionlessclient.services;
-
-import com.exceptionless.exceptionlessclient.models.services.EnvironmentInfo;
-
-public interface EnvironmentInfoCollectorIF {
-  EnvironmentInfo getEnvironmentInfo(EnvironmentInfoGetArgs args);
-}
