--- src/main/java/com/exceptionless/exceptionlessclient/services/RequestInfoCollectorIF.java
@@ -1,9 +1,0 @@
-package com.exceptionless.exceptionlessclient.services;
-
-import com.exceptionless.exceptionlessclient.models.services.RequestInfo;
-
-import java.net.http.HttpRequest;
-
-public interface RequestInfoCollectorIF {
-  RequestInfo getRequestInfo(HttpRequest request, RequestInfoGetArgs args);
-}
