+++ src/test/java/com/exceptionless/exceptionlessclient/plugins/preconfigured/EnvironmentInfoPluginTest.java
@@ -2,8 +2,10 @@ package com.exceptionless.exceptionlessclient.plugins.preconfigured;
 
 import com.exceptionless.exceptionlessclient.TestFixtures;
 import com.exceptionless.exceptionlessclient.configuration.ConfigurationManager;
+import com.exceptionless.exceptionlessclient.configuration.PrivateInformationInclusions;
 import com.exceptionless.exceptionlessclient.models.Event;
 import com.exceptionless.exceptionlessclient.models.EventPluginContext;
+import com.exceptionless.exceptionlessclient.models.services.EnvironmentInfo;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
@@ -22,11 +24,35 @@ public class EnvironmentInfoPluginTest {
   }
 
   @Test
-  public void itCanFillEnvironmentInfo() {
+  public void itShouldNotIncludeMachineNameAndIpAddressUntilExplicitlyTold() {
     assertThat(context.getEvent().getEnvironmentInfo()).isEmpty();
 
-    plugin.run(context,configurationManager);
+    plugin.run(context, configurationManager);
 
     assertThat(context.getEvent().getEnvironmentInfo()).isPresent();
+    EnvironmentInfo info = context.getEvent().getEnvironmentInfo().get();
+    assertThat(info.getMachineName()).isNull();
+    assertThat(info.getIpAddress()).isNull();
+
+    assertThat(info.getProcessorCount()).isNotNull();
+  }
+
+  @Test
+  public void itCanIncludeMachineNameAndIpAddress() {
+    assertThat(context.getEvent().getEnvironmentInfo()).isEmpty();
+    PrivateInformationInclusions inclusions =
+        configurationManager.getPrivateInformationInclusions();
+    inclusions.setIpAddress(true);
+    inclusions.setMachineName(true);
+
+    plugin.run(context, configurationManager);
+
+    assertThat(context.getEvent().getEnvironmentInfo()).isPresent();
+    EnvironmentInfo info = context.getEvent().getEnvironmentInfo().get();
+
+    assertThat(info.getMachineName()).isNotNull();
+    assertThat(info.getIpAddress()).isNotNull();
+
+    assertThat(info.getProcessorCount()).isNotNull();
   }
 }
